refactor(validation): split REGISTRY into per-domain modules (no behavior change)
Convert the monolithic src/rigplane/validation/registry.py into the
rigplane.validation.registry package so future coverage-expansion tasks
(T7-T11 of the harness epic) can edit disjoint per-domain files without
merge conflicts (MOR-637, harness epic T0).

Layout:
- _types.py      — CheckKind, ValueRule, VALUE_RULES, CheckSpec
- _structural.py — checks 1-4 (discovery, freq, mode)
- _levels.py     — checks 5-9 (filter_width, rf_gain, af_level, preamp, attenuator)
- _dsp.py        — checks 10-13 (notch, nb, nr, agc)
- _tuning.py     — checks 14-16 (rit, xit, squelch)
- _surfaces.py   — checks 17-19 (audio, scope, meters)
- _tx.py         — checks 20-21 (tuner, tx.ptt)
- _assembly.py   — order-preserving REGISTRY concat, REGISTRY_BY_ID,
                   get_spec, import-time _validate_registry guard
- _builders.py   — build_template_from_capabilities,
                   build_hamlib_template_from_capabilities
- __init__.py    — facade re-exporting the exact former public API

The assembled REGISTRY is repr-identical (content and order) to the
pre-split tuple — verified by dumping repr() of all 21 specs before and
after the split and diffing (byte-identical). Order matters because the
template generators stable-sort by level and preserve registry order
within each level; the concatenation reproduces positions 1-21 exactly.

Adds a guard test freezing the 21-element check_id set so no check can
be silently lost, duplicated, or renamed by the split.

No logic changes anywhere; every existing
`from rigplane.validation.registry import X` keeps working.
